frontend/requirements.txt
streamlit==1.51.0
Pillow==12.0.0
python-dotenv==1.2.1